Move has_option to AnsiblePlugin
Use a doc fragment for vars plugins

use vars_ sections

boilerplate
